Add support for API tokens
Case PIG-2786: Added an 'api_token' option to the constructor so that a user
will have an explicit place to use their API token. Updated the constructor
unit test to test the new option.

Created a new unit test to test authentication using an API token.

Updated the POD documentation for authentication.

Additionally, updated the TFA unit test since it was failing.